+++ plugin-api/src/main/java/org/sonar/api/batch/sensor/issue/IssueResolution.java
@@ -1,0 +1,45 @@
+/*
+ * Sonar Plugin API
+ * Copyright (C) 2009-2025 SonarSource Sàrl
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package org.sonar.api.batch.sensor.issue;
+
+import java.util.Set;
+import org.sonar.api.batch.fs.InputFile;
+import org.sonar.api.batch.fs.TextRange;
+import org.sonar.api.rule.RuleKey;
+
+/**
+ * Represents a request to resolve one or more issues at a given location.
+ *
+ * @since 13.5
+ */
+public interface IssueResolution {
+
+  enum Status {DEFAULT, FALSE_POSITIVE}
+
+  InputFile inputFile();
+
+  TextRange textRange();
+
+  Status status();
+
+  Set<RuleKey> ruleKeys();
+
+  String comment();
+}